// src/api/openapi/models/admin-matching-create-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zeroone API v1
 * v1
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface AdminMatchingCreateRequest {
    /**
     * 매칭을 요청하는 회원 ID
     */
    'memberId': number;
    /**
     * 매칭될 파트너 회원 ID
     */
    'partnerId': number;
    /**
     * 매칭 상태 (PENDING, RES_ACPT, RES_AUTO, RES_REJ, AUTO, DONE, CANCEL 등)
     */
    'status': AdminMatchingCreateRequestStatusEnum;
    /**
     * 매칭 종류 (AUTO, MANUAL)
     */
    'type': AdminMatchingCreateRequestTypeEnum;
    /**
     * 관리자 메모
     */
    'content'?: string;
    /**
     * 주간 식별자
     */
    'weeklyPeriodIdentifier': string;
}

export const AdminMatchingCreateRequestStatusEnum = {
    Pending: 'PENDING',
    ResAcpt: 'RES_ACPT',
    ResAuto: 'RES_AUTO',
    ResRej: 'RES_REJ',
    Auto: 'AUTO',
    Done: 'DONE',
    Cancel: 'CANCEL'
} as const;

export type AdminMatchingCreateRequestStatusEnum = typeof AdminMatchingCreateRequestStatusEnum[keyof typeof AdminMatchingCreateRequestStatusEnum];
export const AdminMatchingCreateRequestTypeEnum = {
    Auto: 'AUTO',
    Manual: 'MANUAL'
} as const;

export type AdminMatchingCreateRequestTypeEnum = typeof AdminMatchingCreateRequestTypeEnum[keyof typeof AdminMatchingCreateRequestTypeEnum];