// src/api/openapi/models/reset-weekly-matching-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zeroone API
 * Class APIs are served under /api/v5 while legacy domains may still remain on /api/v1 during migration.
 *
 * The version of the OpenAPI document: v5-class-migration
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 주차 매칭 데이터 초기화 요청
 */
export interface ResetWeeklyMatchingRequest {
    /**
     * 초기화할 주차의 식별자 (해당 주의 월요일 날짜)
     */
    'weeklyPeriodIdentifier': string;
}